Node.js Weather App files.
The weather app makes use of the yargs, require node packages, to display the current temperature of the address, entered via the command line. The app is purely based on Node.js and is part of the course: node.js developer course 2, on www.udemy.com.
We make use of the Google API to fetch the coordinates of the address entered by the user. The temperature is fetched through https://darksky.net.
app.js is the main file to run. It reads the address from the command line and calls geocode.js to get the latitude and longitude of the address (via google API).
app.js then calls appWeather which in turn calls tempFetch, to get the current and apparent temperature of the returned latitude and longitude.